+++ backend/config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  secretKey: 'ef4w3fgtuwqw3543tthu6'
+};
